Add kits for SCX version v1.6.3-205
This commit is to "seed" this repository with an SCX version
that should resolve the OMS installation issues, and allow the
OMS team to move forward with changes to their kit building
process.

The "final", tested supported kit for OMS (and Azure) purposes
will be checked in at a later date.